cập nhật lại đường dẫn, đồng nhất folder tránh nhầm lẫn
