+++ front_end/src/components/AppLayout.jsx
@@ -1,6 +1,6 @@
 import '../assets/styles/App.css';
-import Header from '../component/Header';
-import Footer from '../component/Footer';
+import Header from './Header';
+import Footer from './Footer';
 import AppRoutes from '../routes/AppRoutes.jsx';
 import { BrowserRouter as Router } from "react-router-dom";
 import { useLocation } from 'react-router-dom';
