+++ front_end/src/components/Footer.jsx
