+++ front_end/src/components/ProductCard.jsx
@@ -1,59 +1,41 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-import { FiHeart } from 'react-icons/fi';
-import { useWishlist } from '../context/WishlistContext';
+import React, { useState } from "react";
+import { FiShoppingCart, FiHeart } from "react-icons/fi";
+import { useNavigate } from "react-router-dom";
 
 const ProductCard = ({ product }) => {
-  const { addToWishlist, removeFromWishlist, isInWishlist } = useWishlist();
-  const isWishlisted = isInWishlist(product.id);
-
-  const handleWishlistClick = (e) => {
-    e.preventDefault();
-    if (isWishlisted) {
-      removeFromWishlist(product.id);
-    } else {
-      addToWishlist(product);
-    }
+  const [isWishlisted, setIsWishlisted] = useState(false);
+  const navigate = useNavigate();
+  const handleClick = () => {
+    navigate('/product_detail');
   };
 
-  // Lấy ảnh đầu tiên của sản phẩm từ danh sách imageUrls
-  const productImage = product.imageUrls && product.imageUrls.length > 0 
-    ? product.imageUrls[0] 
-    : '/src/assets/images/placeholder.jpg';
+  const addToWishlist = (e) => {
+    e.stopPropagation();
+    setIsWishlisted(!isWishlisted);
+  };
 
   return (
-    <Link to={`/product/${product.id}`} className="group">
-      <div className="relative overflow-hidden rounded-lg bg-white shadow-md transition-all hover:shadow-lg">
-        <div className="aspect-w-1 aspect-h-1 w-full overflow-hidden">
-          <img
-            src={productImage}
-            alt={product.name}
-            className="h-full w-full object-cover object-center transition-transform duration-300 group-hover:scale-105"
-          />
+    <div className="flex flex-col h-[430px] bg-white rounded-xl shadow p-0 cursor-pointer" onClick={handleClick}>
+      <div className="w-full aspect-square bg-gray-50 flex items-center justify-center overflow-hidden">
+        <img src={product.image} alt={product.name} className="object-cover w-full h-full rounded-lg" />
+      </div>
+      <div className="flex flex-col flex-1 justify-between p-4">
+        <div>
+          <div className="font-semibold text-lg line-clamp-2 min-h-[3rem]">{product.name}</div>
+          <div className="text-gray-500 text-sm mb-2">{product.category}</div>
+          <div className="font-bold text-red-500 text-xl mb-2">{product.price?.toLocaleString()} đ</div>
         </div>
-        <div className="p-4">
-          <h3 className="mb-2 text-lg font-semibold text-gray-900">{product.name}</h3>
-          <p className="mb-2 text-sm text-gray-600">{product.categoryName}</p>
-          <div className="flex items-center justify-between">
-            <p className="text-lg font-bold text-primary">
-              {new Intl.NumberFormat('vi-VN', {
-                style: 'currency',
-                currency: 'VND'
-              }).format(product.price)}
-            </p>
-            <button
-              onClick={handleWishlistClick}
-              className={`rounded-full p-2 transition-colors ${
-                isWishlisted ? 'text-red-500' : 'text-gray-400 hover:text-red-500'
-              }`}
-            >
-              <FiHeart className="h-5 w-5" />
-            </button>
-          </div>
+        <div className="flex items-center justify-between mt-4">
+          <button className="flex-1 bg-red-500 text-white py-2 rounded mr-2 hover:bg-red-600 flex items-center justify-center">
+            <FiShoppingCart className="mr-2" /> Thêm vào giỏ
+          </button>
+          <button className={`p-2 rounded-full border ${isWishlisted ? 'bg-red-100 text-red-500' : 'text-gray-400 hover:text-red-500'}`} onClick={addToWishlist}>
+            <FiHeart />
+          </button>
         </div>
       </div>
-    </Link>
+    </div>
   );
 };
 
-export default ProductCard; 
+export default ProductCard;
